# Accelerate `minimum(A; dims = 1)` for cartesian indexed cases.