Our `mapreduce` use `mapreduce_impl` for linear-indexable inputs to accelerate min/max related calls, which is critical to Float cases especially when dim1 is included in reduction. 
But the optimization is missing for cartesian cases.

This PR try to fuse more `mapreduce_impl` for free speed:
`minimum` Benchmark (combined with #43604, which widen the optimizaiton to much shorten length):
```julia
using Random, BenchmarkTools                   
Random.seed!(0);
a = view(rand(128,128),1:128,1:128);# PR 43604      This PR
@btime minimum($a, dims = 1);       #28.100 μs  --> 9.400 μs 
@btime minimum($a, dims = 2);       #4.071 μs   --> 4.086 μs 
@btime minimum($a, dims = (1,2));   #26.700 μs  --> 9.700 μs
@btime minimum($a);                 #30.400 μs  --> 9.100 μs
```